% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReference.R
\name{buildReference}
\alias{buildReference}
\title{Function for building a Symphony reference starting from expression matrix}
\usage{
buildReference(
  exp_ref,
  metadata_ref,
  vars = NULL,
  K = 50,
  verbose = FALSE,
  do_umap = TRUE,
  do_normalize = TRUE,
  vargenes_method = "vst",
  topn = 2000,
  tau = 0,
  theta = 2,
  save_uwot_path = NULL,
  d = 20,
  additional_genes = NULL
)
}
\arguments{
\item{exp_ref}{Reference gene expression (genes by cells)}

\item{metadata_ref}{Reference cell metadata (cells by attributes)}

\item{vars}{Reference variables to Harmonize over e.g. c('donor', 'technology')}

\item{K}{Number of centroids}

\item{verbose}{Verbose output}

\item{do_umap}{Perform UMAP visualization on harmonized reference embedding}

\item{do_normalize}{Perform log(CP10K) normalization}

\item{vargenes_method}{Variable gene selection method (either 'vst' or 'mvp')}

\item{topn}{Number of variable genes to subset by}

\item{tau}{Tau parameter for Harmony step}

\item{theta}{Theta parameter(s) for Harmony step}

\item{save_uwot_path}{Absolute path to save the uwot model (if do_umap is TRUE)}

\item{d}{Number of PC dimensions}

\item{additional_genes}{Any custom genes (e.g. marker genes) to include in addition to variable genes}
}
\description{
Function for building a Symphony reference starting from expression matrix
}
