% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knncorr.R
\name{calcknncorrWithinQueryBatch}
\alias{calcknncorrWithinQueryBatch}
\title{Non-exported (called by calcknncorrWithinQuery above)
Calculates the k-NN correlation within a query batch.}
\usage{
calcknncorrWithinQueryBatch(Z_pca, Z_mapping, k, distance)
}
\arguments{
\item{Z_pca}{PCA embedding}

\item{Z_mapping}{Reference mapping embedding}

\item{k}{Number of neighbors to use for kNN-correlation calculation}

\item{distance}{either 'euclidean' or 'cosine'}
}
\description{
Non-exported (called by calcknncorrWithinQuery above)
Calculates the k-NN correlation within a query batch.
}
