% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceScores.R
\name{calcPerClusterConfidence}
\alias{calcPerClusterConfidence}
\title{---- Per-cluster Confidence Score ----
Calculates the Mahalanobis distance from user-defined query clusters to their nearest
reference centroid after initial projection into reference PCA space.
All query cells in a cluster get the same score. Higher distance indicates less confidence.}
\usage{
calcPerClusterConfidence(
  reference,
  query,
  query_cluster_labels,
  metric = "mahalanobis"
)
}
\arguments{
\item{reference}{Reference object as returned by Symphony buildReference()}

\item{query}{Query object as returned by Symphony mapQuery()}

\item{query_cluster_labels}{Vector of user-defined labels denoting clusters / putative novel cell type to calculate the score for}

\item{metric}{Uses Mahalanobis by default, but added as a parameter for potential future use}
}
\description{
---- Per-cluster Confidence Score ----
Calculates the Mahalanobis distance from user-defined query clusters to their nearest
reference centroid after initial projection into reference PCA space.
All query cells in a cluster get the same score. Higher distance indicates less confidence.
}
