% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ref_cache}
\alias{compute_ref_cache}
\title{Computes the Symphony reference compression terms, Nr and C.}
\usage{
compute_ref_cache(Rr, Zr)
}
\arguments{
\item{Rr}{Soft cluster assignments of reference cells (cols) across clusters (rows).}

\item{Zr}{Corrected embedding for reference cells (cols) in harmonized PCs (rows).}
}
\description{
Computes the Symphony reference compression terms, Nr and C.
}
