% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_normalize_cpp}
\alias{cosine_normalize_cpp}
\title{Cosine normalization}
\usage{
cosine_normalize_cpp(V, dim)
}
\arguments{
\item{V}{matrix}

\item{dim}{rows (dim 1) or col (dim 2)}
}
\description{
Cosine normalization
}
