% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate}
\alias{evaluate}
\title{Function for evaluating F1 by cell type,
Adapted from automated cell type identifiaction benchmarking paper (Abdelaal et al. Genome Biology, 2019)}
\usage{
evaluate(true, predicted)
}
\arguments{
\item{true}{vector of true labels}

\item{predicted}{vector of predicted labels}
}
\description{
Function for evaluating F1 by cell type,
Adapted from automated cell type identifiaction benchmarking paper (Abdelaal et al. Genome Biology, 2019)
}
