% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableGenes.R
\name{findVariableGenes}
\alias{findVariableGenes}
\title{Function to find variable genes using mean variance relationship method}
\usage{
findVariableGenes(
  X,
  groups,
  min_expr = 0.1,
  max_expr = Inf,
  min_dispersion = 0,
  max_dispersion = Inf,
  num.bin = 20,
  binning.method = "equal_width",
  return_top_n = 0
)
}
\description{
Function to find variable genes using mean variance relationship method
}
