% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_betas}
\alias{get_betas}
\title{Returns the batch coefficients of the linear mixture model as a 3D tensor.}
\usage{
get_betas(R, Z, lambda, design)
}
\arguments{
\item{R}{Cell cluster assignments}

\item{Z}{PCA embedding}

\item{lambda}{Diagonal matrix for ridge regression}

\item{design}{Design matrix}
}
\description{
Returns the batch coefficients of the linear mixture model as a 3D tensor.
}
