% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moe_correct_ref}
\alias{moe_correct_ref}
\title{Computes the corrected query cell embedding.}
\usage{
moe_correct_ref(Zq, Xq, Rq, Nr, RrZtr)
}
\arguments{
\item{Zq}{Query cells projected into PC space (d x m)}

\item{Xq}{Query design matrix ((c + 1) x m)}

\item{Rq}{Query soft cluster assignments across reference clusters (k x m)}

\item{Nr}{Reference cluster sizes (first compression term) (length k)}

\item{RrZtr}{Second reference compression term (C) (k x d)}
}
\description{
Computes the corrected query cell embedding.
}
