% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcs_exprs_small}
\alias{pbmcs_exprs_small}
\title{Log(CP10k+1) normalized counts matrix (genes by cells) for 10x PBMCs dataset for vignette.}
\format{
: Sparse matrix (dgCMatrix): dimensions 1,764 genes by 1,200 cells
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcs_exprs_small
}
\description{
Log(CP10k+1) normalized counts matrix (genes by cells) for 10x PBMCs dataset for vignette.
}
\keyword{datasets}
