% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcs_meta_small}
\alias{pbmcs_meta_small}
\title{Metadata for 10x PBMCs dataset for vignette.}
\format{
: A data frame with 1,200 cells and 7 metadata fields.
\describe{
\item{cell_id}{unique cell ID}
\item{donor}{dataset (3pv1, 3pv2, or 5p)}
\item{nUMI}{number of UMIs}
\item{nGene}{number of genes}
\item{percent_mito}{percent mito genes}
\item{cell_type}{cell type assigned in Symphony publication}
\item{cell_type_broad}{cell subtype assigned in Symphony publication}

}
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcs_meta_small
}
\description{
Metadata for 10x PBMCs dataset for vignette.
}
\keyword{datasets}
