% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowSDs}
\alias{rowSDs}
\title{Calculate standard deviations by row}
\usage{
rowSDs(A, row_means = NULL, weights = NULL)
}
\arguments{
\item{A}{expression matrix (genes by cells)}

\item{row_means}{row means}

\item{weights}{weights for weighted standard dev calculation}
}
\value{
A vector of row standard deviations
}
\description{
Calculate standard deviations by row
}
