% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knncorr.R
\name{runPCAQueryAlone}
\alias{runPCAQueryAlone}
\title{Runs a standard PCA pipeline on query (1 batch). Assumes query_exp is already normalized.}
\usage{
runPCAQueryAlone(query_exp, topn = 2000, d = 20, seed = 1)
}
\arguments{
\item{query_exp}{Query expression matrix (genes x cells)}

\item{topn}{Number of variable genes to use}

\item{d}{Number of dimensions}

\item{seed}{random seed}
}
\value{
A matrix of PCs by cells
}
\description{
Runs a standard PCA pipeline on query (1 batch). Assumes query_exp is already normalized.
}
