% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleDataWithStats}
\alias{scaleDataWithStats}
\title{Scale data with given mean and standard deviations}
\usage{
scaleDataWithStats(A, mean_vec, sd_vec, margin = 1, thresh = 10)
}
\arguments{
\item{A}{expression matrix (genes by cells)}

\item{mean_vec}{vector of mean values}

\item{sd_vec}{vector of standard deviation values}

\item{margin}{1 for row-wise calculation}

\item{thresh}{threshold to clip max values}
}
\value{
A matrix of scaled expression values.
}
\description{
Scale data with given mean and standard deviations
}
