% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_cluster}
\alias{soft_cluster}
\title{Computes the soft cluster assignments of query cells across reference clusters.}
\usage{
soft_cluster(Y, Z, sigma)
}
\arguments{
\item{Y}{Reference cluster centroid locations}

\item{Z}{Query cells projected into PC space (d x m)}

\item{sigma}{Soft k-means "fuzziness" parameter, sigma = 0 is hard clustering}
}
\description{
Computes the soft cluster assignments of query cells across reference clusters.
}
