% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symphony-package.R
\docType{package}
\name{symphony-package}
\alias{symphony}
\alias{symphony-package}
\title{symphony: Efficient and Precise Single-Cell Reference Atlas Mapping}
\description{
Implements the Symphony single-cell reference building and query mapping algorithms and additional functions described in Kang et al \url{https://www.nature.com/articles/s41467-021-25957-x}.
}
\author{
\strong{Maintainer}: Joyce Kang \email{joyce.b.kang@gmail.com} (\href{https://orcid.org/0000-0002-1962-1291}{ORCID})

Authors:
\itemize{
  \item Ilya Korsunsky \email{ilya.korsunsky@gmail.com} (\href{https://orcid.org/0000-0003-4848-3948}{ORCID})
  \item Soumya Raychaudhuri (\href{https://orcid.org/0000-0002-1901-8265}{ORCID})
}

}
\keyword{internal}
