% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableGenes.R
\name{vargenes_vst}
\alias{vargenes_vst}
\title{Function to find variable genes using variance stabilizing transform (VST) method}
\usage{
vargenes_vst(object, groups, topn, loess.span = 0.3)
}
\description{
Function to find variable genes using variance stabilizing transform (VST) method
}
