% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceFromHarmonyObj.R
\name{buildReferenceFromHarmonyObj}
\alias{buildReferenceFromHarmonyObj}
\title{Function for building a Symphony reference from a Harmony object. Useful if you would like your
code to be more modular. Note that you must have your saved vargenes_means_sds and PCA loadings.}
\usage{
buildReferenceFromHarmonyObj(
  harmony_obj,
  metadata,
  vargenes_means_sds,
  pca_loadings,
  verbose = TRUE,
  do_umap = TRUE,
  save_uwot_path = NULL
)
}
\arguments{
\item{harmony_obj}{Harmony object (output from HarmonyMatrix())}

\item{metadata}{Reference cell metadata (cells by attributes)}

\item{vargenes_means_sds}{Variable genes in dataframe with columns ('symbol', 'mean', 'stddev')}

\item{pca_loadings}{Gene loadings from PCA (= irlba(ref_exp_scaled, nv = 20)$u)}

\item{verbose}{Verbose output}

\item{do_umap}{Perform UMAP visualization on harmonized reference embedding}

\item{save_uwot_path}{Absolute path to save the uwot model (if do_umap is TRUE)}
}
\description{
Function for building a Symphony reference from a Harmony object. Useful if you would like your
code to be more modular. Note that you must have your saved vargenes_means_sds and PCA loadings.
}
