% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate}
\alias{evaluate}
\title{Function for evaluating F1 by cell type,
adapted from automated cell type identifiaction benchmarking paper (Abdelaal et al. Genome Biology, 2019)}
\usage{
evaluate(true, predicted)
}
\arguments{
\item{true}{vector of true labels}

\item{predicted}{vector of predicted labels}
}
\value{
A list of results with confusion matrix ($Conf), median F1-score ($MedF1), F1 scores per class ($F1), and accuracy ($Acc).
}
\description{
Function for evaluating F1 by cell type,
adapted from automated cell type identifiaction benchmarking paper (Abdelaal et al. Genome Biology, 2019)
}
