% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableGenes.R
\name{findVariableGenes}
\alias{findVariableGenes}
\title{Function to find variable genes using mean variance relationship method}
\usage{
findVariableGenes(
  X,
  groups,
  min_expr = 0.1,
  max_expr = Inf,
  min_dispersion = 0,
  max_dispersion = Inf,
  num.bin = 20,
  binning.method = "equal_width",
  return_top_n = 0
)
}
\arguments{
\item{X}{expression matrix}

\item{groups}{vector of groups}

\item{min_expr}{min expression cutoff}

\item{max_expr}{max expression cutoff}

\item{min_dispersion}{min dispersion cutoff}

\item{max_dispersion}{max dispersion cutoff}

\item{num.bin}{number of bins to use for scaled analysis}

\item{binning.method}{how bins are computed}

\item{return_top_n}{returns top n genes}
}
\value{
A data.frame of variable genes
}
\description{
Function to find variable genes using mean variance relationship method
}
