% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knnPredict}
\alias{knnPredict}
\title{Predict cell annotations using knn method}
\usage{
knnPredict(query_obj, ref_obj, train_labels, k = 5)
}
\arguments{
\item{query_obj}{Query object}

\item{ref_obj}{Reference object}

\item{train_labels}{vector of labels to train}

\item{k}{num neighbors}
}
\description{
Predict cell annotations using knn method
}
