% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableGenes.R
\name{vargenes_vst}
\alias{vargenes_vst}
\title{Function to find variable genes using variance stabilizing transform (vst) method}
\usage{
vargenes_vst(object, groups, topn, loess.span = 0.3)
}
\arguments{
\item{object}{expression matrix}

\item{groups}{finds variable genes within each group then pools}

\item{topn}{Return top n genes}

\item{loess.span}{Loess span parameter used when fitting the variance-mean relationship}
}
\value{
A data.frame of variable genes, with means and standard deviations.
}
\description{
Function to find variable genes using variance stabilizing transform (vst) method
}
